Refactor to focus on relay server
* Move clients to examples
* Seperate server executable from library
* New module for relay_session
* Remove some duplicate code
