+++ RelayProofsOfConcept/EddsaTokioServer/.gitignore
@@ -1,5 +1,6 @@
 # Generated by Cargo
 # will have compiled files and executables
+*/target/
 /target/
 common/target/
 
@@ -13,3 +14,6 @@ Cargo.lock
 .idea
 .DS_Store
 .*.swp
+
+keys*
+signature
