+++ RelayProofsOfConcept/EddsaTokioServer/Cargo.toml
@@ -10,13 +10,15 @@ serde_json = "1.0"
 futures = "0.1"
 tokio-core = "0.1"
 tokio-io = "0.1"
+tokio = "0.1"
+tokio-codec = "0.1"
 byteorder = "1.0"
 dict = "0.1.5"
 chrono = "0.4"
 hex = "0.3.2"
 structopt = "0.2"
 
-relay-server-common = { path = "../EddsaTokioServer/common" }
+relay-server-common = { path = "../EddsaTokioServer/relay-server-common" }
 
 multi-party-ed25519 = { git = "https://github.com/KZen-networks/multi-party-eddsa" }
