+++ RelayProofsOfConcept/EddsaTokioServer/examples/eddsa_key_gen_client.rs
