+++ RelayProofsOfConcept/EddsaTokioServer/sign.sh
@@ -12,7 +12,7 @@ cargo run --package relay-server --bin server -- -P $n&
 sleep 2
 for i in $(seq 1 $n);
 do
-    cargo run --bin eddsa_sign_client -- 127.0.0.1:8080 "keys$i" -P $n $1 &
+    cargo run --example eddsa_sign_client -- 127.0.0.1:8080 "keys$i" -P $n $1 &
     sleep 1
 done
